+++ backend/src/baserow/api/jobs/serializers.py
@@ -36,14 +36,7 @@ class JobSerializer(serializers.ModelSerializer):
 
     @extend_schema_field(OpenApiTypes.STR)
     def get_type(self, instance):
-        # It could be that the field related to the instance is already in the context
-        # else we can call the specific_class property to find it.
-        field = self.context.get("instance_type")
-
-        if not field:
-            field = job_type_registry.get_by_model(instance.specific_class)
-
-        return field.type
+        return job_type_registry.get_by_model(instance.specific_class).type
 
 
 class CreateJobSerializer(serializers.Serializer):