+++ backend/src/baserow/contrib/database/file_import/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 from baserow.contrib.database.models import Database
 from baserow.contrib.database.table.models import Table
-from baserow.core.jobs.mixins import JobWithUserDataMixin
+from baserow.core.jobs.mixins import JobWithUndoRedoIds, JobWithWebsocketId
 from baserow.core.jobs.models import Job
 
 
@@ -20,8 +20,7 @@ def default_report():
     return {"failing_rows": {}}
 
 
-class FileImportJob(JobWithUserDataMixin, Job):
-    user_data_to_save = ["user_websocket_id"]
+class FileImportJob(JobWithWebsocketId, JobWithUndoRedoIds, Job):
 
     database = models.ForeignKey(
         Database,